refactor(cli): migrate sentinel_hub tooling onto discovery
Move sentinel_hub's refresher/writer/curated_ids/prober/validator to
earthlens.sentinel_hub.cli (the SDK DataCollection lister + offline evalscript lint),
dropping core's last direct earthlens.sentinel_hub imports. Relocate the CLI tests.
Shrink both guard allow-lists (sentinel_hub).
